kernels: int64 bases in every expert/row stride product; straddle the 2^31 boundary in every carrier (#87)
The expert-id cast the issue asked for shipped in 0.13.2 (NF4) and 0.14.0
(MXFP4). Two carriers of the same pattern were still int32: the four
fp8_paged_attn decode kernels scaled the block-table row by k_row_bytes /
v_row_bytes in int32 (the fp8_kv writer already widened, so a pool past 2^31
bytes was written right and read from a wrapped offset), and the M-tile
kernels' row0 (activation / output row offsets, wrapping at T * max(K, N)
>= 2^31 elements; _gemm_int4_b32_grouped already widened it).

kernel/test_expert_offset_boundary.py observes what the issue flagged by
inspection: for the NF4 decode GEMV (scalar, split-K, wide-load, vec-load,
dot-pad, dot-pad split-K -- each route asserted through the dispatch tally),
the NF4 M-tile and dgrad, the MXFP4 GEMM / GEMV / gemv_mxfp4_b32, the int4-b32
GEMV and M-tile, and the fp8 paged decode kernels in all four modes, the
experts (or pool rows) whose base offsets sit just below and just above 2^31
are compared with the pure-torch references, each above-boundary case in its
own subprocess (an illegal access poisons the CUDA context). Geometry: N=256,
K=1024 (2^17 B per expert; experts 16383 | 16384; 2.0 GiB + scales), the
dot-pad census shape 1536 x 2048 (experts 1365 | 1366), and a K pool padded
to 1 MiB rows (rows 2047 | 2048). Skips legibly without a CUDA device or
under 4 GiB free; named in ci.yml in its own process.

docs/KERNEL_CONTRACT.md gains a dated Boundaries note with the rule and the
carrier list; CHANGELOG under Unreleased.

+++ kernel/fp8_paged_attn.py
@@ -144,8 +144,12 @@ if _TRITON:
         for start in range(t_lo, t_hi, KTILE):
             tok = start + offs_t
             t_mask = (tok < t_hi) & (tok >= w_lo)
+            # int64 BEFORE the row-stride product: blk * k_row_bytes is the
+            # pool byte offset of a 16-token row and wraps in int32 once the
+            # pool passes 2^31 bytes (the writer, fp8_kv, already widens;
+            # gnf4#87 pattern, docs/KERNEL_CONTRACT.md Boundaries).
             blk = tl.load(table_ptr + b * stride_tb + tok // BT,
-                          mask=t_mask, other=0)
+                          mask=t_mask, other=0).to(tl.int64)
 
             k_base = blk * k_row_bytes
             ku = tl.load(kpool_u8 + k_base[:, None] + pay_off,
@@ -301,7 +305,8 @@ if _TRITON:
         acc = tl.zeros([BLOCK_Q, D], tl.float32)
 
         for blk_i in range(blk_lo, blk_hi):
-            blk = tl.load(table_ptr + b * stride_tb + blk_i)
+            # int64 before the row-stride product (see the split kernel)
+            blk = tl.load(table_ptr + b * stride_tb + blk_i).to(tl.int64)
             t0 = blk_i * BT
             t_mask_c = ((t0 + c_tok) < t_len) & ((t0 + c_tok) >= w_lo)
 
@@ -515,8 +520,12 @@ if _TRITON:
         for start in range(t_lo, t_hi, KTILE):
             tok = start + offs_t
             t_mask = (tok < t_hi) & (tok >= w_lo)
+            # int64 BEFORE the row-stride product: blk * k_row_bytes is the
+            # pool byte offset of a 16-token row and wraps in int32 once the
+            # pool passes 2^31 bytes (the writer, fp8_kv, already widens;
+            # gnf4#87 pattern, docs/KERNEL_CONTRACT.md Boundaries).
             blk = tl.load(table_ptr + b * stride_tb + tok // BT,
-                          mask=t_mask, other=0)
+                          mask=t_mask, other=0).to(tl.int64)
 
             k_base = blk * k_row_bytes
             if LAYOUT_HEADS:
@@ -774,7 +783,8 @@ if _TRITON:
         c_prev = 1.0
 
         for blk_i in range(blk_lo, blk_hi):
-            blk = tl.load(table_ptr + b * stride_tb + blk_i)
+            # int64 before the row-stride product (see the split kernel)
+            blk = tl.load(table_ptr + b * stride_tb + blk_i).to(tl.int64)
             t_mask_c = ((blk_i * BT + c_tok) < t_len) & ((blk_i * BT + c_tok) >= w_lo)
 
             k_base = blk * k_row_bytes
